+++ persistent/test/Database/Persist/THSpec.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingStrategies #-}
@@ -23,6 +24,8 @@
 
 module Database.Persist.THSpec where
 
+import System.Environment
+import Data.Time
 import Control.Applicative (Const(..))
 import Data.Aeson (decode, encode)
 import Data.ByteString.Lazy.Char8 ()
@@ -426,6 +429,54 @@ spec = describe "THSpec" $ do
 
             show (CustomPrimaryKeyKey 0) `shouldBe` "CustomPrimaryKeyKey {unCustomPrimaryKeyKey = 0}"
             read (show (CustomPrimaryKeyKey 0)) `shouldBe` CustomPrimaryKeyKey 0
+
+    describe "tabulateEntityA" $ do
+        it "works" $ do
+            person <-
+                tabulateEntityA $ \case
+                    PersonName ->
+                        pure "Matt"
+                    PersonAge -> do
+                        (year, _, _) <- toGregorian . utctDay <$> getCurrentTime
+                        pure $ Just (fromInteger year - 1988)
+                    PersonFoo -> do
+                        _ <- lookupEnv "PERSON_FOO" :: IO (Maybe String)
+                        pure Bar
+                    PersonAddress ->
+                        pure $ Address  "lol no" "Denver" Nothing
+                    PersonId ->
+                        pure $ toSqlKey 123
+            expectedAge <- fromInteger . subtract 1988 . (\(a, _, _) -> a) . toGregorian . utctDay <$> getCurrentTime
+            person `shouldBe` Entity (toSqlKey 123) Person
+                { personName =
+                    "Matt"
+                , personAge =
+                    Just expectedAge
+                , personFoo =
+                    Bar
+                , personAddress =
+                    Address  "lol no" "Denver" Nothing
+                }
+
+    describe "tabulateEntity" $ do
+        it "works" $ do
+            let
+                addressTabulate =
+                    tabulateEntity $ \case
+                        AddressId ->
+                            toSqlKey 123
+                        AddressStreet ->
+                            "nope"
+                        AddressCity ->
+                            "Denver"
+                        AddressZip ->
+                            Nothing
+            addressTabulate `shouldBe`
+                Entity (toSqlKey 123) Address
+                    { addressStreet = "nope"
+                    , addressCity = "Denver"
+                    , addressZip = Nothing
+                    }
 
 (&) :: a -> (a -> b) -> b
 x & f = f x